Refactor attention e2e tests
This PR creates a common utils directory for
e2e test functions and shapes and splits the
attention tests into multiple files.
