+++ tests/kernel/wave/attention/vanilla_attention_test.py
@@ -1,23 +1,19 @@
-# Copyright 2024 The IREE Authors
+# Copyright 2025 The IREE Authors
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions.
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-import logging
 import pytest
 import torch
 from torch.nn import functional as F
 import math
-import unittest
 import iree.turbine.kernel as tk
 import iree.turbine.kernel.lang as tkl
 import iree.turbine.kernel.wave as tkw
 from iree.turbine.kernel.lang.global_symbols import *
-from iree.turbine.kernel.wave.iree_utils import generate_iree_ref
 from iree.turbine.kernel.wave.utils import (
     get_default_run_config,
-    get_default_arch,
     get_default_scheduling_params,
     get_mfma_load_elems_per_thread,
     get_mfma_store_elems_per_thread,
@@ -25,336 +21,19 @@ from iree.turbine.kernel.wave.utils import (
     device_zeros,
 )
 from iree.turbine.kernel.wave.constraints import MMAType
-from iree.turbine.kernel.wave.templates.decode_attention import (
-    get_decode_attention_kernels,
-)
 import os
-import json
-from torch.testing import assert_close, assert_allclose
-from enum import Enum
-
-
-require_e2e = pytest.mark.require_e2e
-require_cdna3 = pytest.mark.skipif(
-    "gfx94" not in get_default_arch(), reason="Default device is not CDNA3"
+from torch.testing import assert_allclose
+from ..common.utils import (
+    require_e2e,
+    require_cdna3,
+    enable_scheduling_barriers,
+    dump_generated_mlir,
 )
-# Whether to dump the generated MLIR module.
-test_dump_generated_mlir = int(os.environ.get("WAVE_DUMP_MLIR", 0))
-# Whether to use scheduling group barriers (needs LLVM fix).
-enable_scheduling_barriers = int(os.environ.get("WAVE_USE_SCHED_BARRIERS", 0))
-
-# Add test shapes for validation and performance testing.
-perf_test = lambda *a: pytest.param(*a, marks=pytest.mark.perf_only)
-default_test_shapes = {}
-# Order of shapes: (B, M, N, K1, K2)
-default_test_shapes["test_attention"] = [
-    (8, 128, 128, 64, 256),
-    (40, 1024, 64, 64, 1024),
-]
-default_test_shapes["test_attention"] += [
-    perf_test(x) for x in default_test_shapes["test_attention"]
-]
-
-
-user_specified_test_shapes = ""
-
-test_params_path = os.environ.get("TEST_PARAMS_PATH", None)
-
-if test_params_path:
-    with open(test_params_path, "r") as file:
-        user_specified_test_shapes = json.load(file)
-
-
-def get_test_shapes(test_name: str) -> list[tuple[int]]:
-    if test_name in user_specified_test_shapes:
-        return user_specified_test_shapes[test_name]
-    return default_test_shapes[test_name]
+from ..common.shapes import get_test_shapes
 
 
 @require_e2e
-@pytest.mark.parametrize("shape", get_test_shapes("test_attention"))
-@pytest.mark.parametrize("enable_scheduling", [False])
-@pytest.mark.parametrize(
-    "mfma_variant",
-    [
-        MMAType.F32_16x16x16_F16,
-        MMAType.F32_32x32x8_F16,
-    ],
-)
-def testChainedGemm(
-    shape: tuple[int],
-    enable_scheduling: bool,
-    mfma_variant: MMAType,
-    request,
-):
-    run_bench = request.config.getoption("--runperf")
-    dump_perf = request.config.getoption("--dump-perf-files-path")
-    # Input sizes
-    B = tkl.sym.B
-    M = tkl.sym.M
-    N = tkl.sym.N
-    K1 = tkl.sym.K1
-    K2 = tkl.sym.K2
-    # Workgroup tile sizes
-    BLOCK_B = tkl.sym.BLOCK_B
-    BLOCK_M = tkl.sym.BLOCK_M
-    BLOCK_N = tkl.sym.BLOCK_N
-    BLOCK_K2 = tkl.sym.BLOCK_K2
-    # Address space (for GPU, shared(1) or global(0))
-    ADDRESS_SPACE = tkl.sym.ADDRESS_SPACE
-    # Other hyperparameters
-    LOAD_ELEMS_PER_THREAD = tkl.sym.LOAD_ELEMS_PER_THREAD
-    STORE_ELEMS_PER_THREAD = tkl.sym.STORE_ELEMS_PER_THREAD
-
-    # Expose user-constraints
-    constraints: list[tkw.Constraint] = [tkw.WorkgroupConstraint(M, BLOCK_M, 0)]
-    constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 1)]
-    constraints += [tkw.WorkgroupConstraint(B, BLOCK_B, 2)]
-    constraints += [tkw.TilingConstraint(K2, BLOCK_K2)]
-    constraints += [tkw.WaveConstraint(M, BLOCK_M / 2)]
-    constraints += [tkw.WaveConstraint(N, BLOCK_N / 2)]
-
-    constraints += [
-        tkw.HardwareConstraint(
-            threads_per_wave=64,
-            waves_per_block=(2, 2, 1),
-            mma_type=mfma_variant,
-            vector_shapes={B: 0},
-        )
-    ]
-
-    i = tkw.IndexMapping.iterator(0)
-    j = tkw.IndexMapping.iterator(1)
-    k = tkw.IndexMapping.iterator(2)
-    mapping = tkw.IndexMapping(
-        num_iterators=3, inputs={B: i, M: j, N: k}, outputs={B: i, N: k, M: j}
-    )
-
-    @tkw.wave(constraints)
-    def chained_gemm(
-        q: tkl.Memory[B, M, K1, GLOBAL_ADDRESS_SPACE, tkl.f16],
-        k: tkl.Memory[B, K2, K1, ADDRESS_SPACE, tkl.f16],
-        v: tkl.Memory[B, N, K2, ADDRESS_SPACE, tkl.f16],
-        c: tkl.Memory[B, N, M, GLOBAL_ADDRESS_SPACE, tkl.f32],
-    ):
-        c_reg = tkl.Register[B, M, N, tkl.f32](0.0)
-
-        @tkw.reduction(K2, init_args=[c_reg])
-        def repeat(
-            acc: tkl.Register[B, M, N, tkl.f32]
-        ) -> tkl.Register[B, M, N, tkl.f32]:
-            inner_acc = tkl.Register[B, K2, M, tkl.f32](0.0)
-            q_reg = tkw.read(q, elements_per_thread=LOAD_ELEMS_PER_THREAD)
-            k_reg = tkw.read(k, elements_per_thread=LOAD_ELEMS_PER_THREAD)
-            kq_reg = tkw.mma(k_reg, q_reg, inner_acc)
-            qk_reg = tkw.permute(kq_reg, target_shape=[B, M, K2])
-            qk_cast_reg = tkw.cast(qk_reg, tkl.f16)
-            v_reg = tkw.read(v, elements_per_thread=LOAD_ELEMS_PER_THREAD)
-            acc = tkw.mma(qk_cast_reg, v_reg, acc)
-            return acc
-
-        # repeat represents the results of the loop
-        tkw.write(
-            repeat, c, mapping=mapping, elements_per_thread=STORE_ELEMS_PER_THREAD
-        )
-
-    hyperparams = {
-        ADDRESS_SPACE: SHARED_ADDRESS_SPACE,
-        LOAD_ELEMS_PER_THREAD: get_mfma_load_elems_per_thread(mfma_variant),
-        STORE_ELEMS_PER_THREAD: get_mfma_store_elems_per_thread(mfma_variant),
-        BLOCK_B: 1,
-        BLOCK_M: 64,
-        BLOCK_N: 64,
-        BLOCK_K2: 32,
-        B: shape[0],
-        M: shape[1],
-        N: shape[2],
-        K1: shape[3],
-        K2: shape[4],
-    }
-    hyperparams.update(get_default_scheduling_params())
-    config = get_default_run_config()
-    if run_bench:
-        config["benchmark_batch_size"] = 10
-        config["benchmark_repetitions"] = 3
-    if dump_perf is not None:
-        perf_filename = request.node.name + ".json"
-        config["benchmark_results_file"] = os.path.join(
-            dump_perf, "tk_" + perf_filename
-        )
-
-    with tk.gen.TestLaunchContext(
-        hyperparams,
-        canonicalize=True,
-        run=True,
-        run_bench=run_bench,
-        run_config=config,
-        schedule=enable_scheduling,
-        use_scheduling_barriers=enable_scheduling_barriers,
-    ):
-        q = device_randn(shape[0], shape[1], shape[3], dtype=torch.float16)
-        k = device_randn(shape[0], shape[4], shape[3], dtype=torch.float16)
-        v = device_randn(shape[0], shape[2], shape[4], dtype=torch.float16)
-        output = device_zeros(shape[0], shape[2], shape[1], dtype=torch.float32)
-        mb = chained_gemm(q, k, v, output)
-
-        if test_dump_generated_mlir:
-            filename = f"wave_cgemm_{'x'.join(map(str, shape))}.mlir"
-            with open(filename, "w") as f:
-                f.write(mb.module_op.get_asm())
-
-        iree_ref = torch.zeros(shape[0], shape[2], shape[1], dtype=torch.float32)
-        generate_iree_ref(
-            "chain_mmt", [q, k, v], [iree_ref], config, run_bench=run_bench
-        )
-        assert_close(output, iree_ref, check_device=False)
-
-
-# This test requires some more analysis on the index sequences between
-# the two chained GEMMs.
-@require_e2e
-@require_cdna3
-@pytest.mark.parametrize("shape", get_test_shapes("test_attention"))
-@pytest.mark.parametrize("enable_scheduling", [False])
-@pytest.mark.parametrize(
-    "mfma_variant",
-    [
-        (MMAType.F32_32x32x16_F8, MMAType.F32_32x32x16_K4_F8),
-        (MMAType.F32_16x16x32_F8, MMAType.F32_16x16x32_K4_F8),
-    ],
-)
-def testChainedGemmF8(
-    shape: tuple[int], enable_scheduling: bool, mfma_variant: tuple[MMAType], request
-):
-    run_bench = request.config.getoption("--runperf")
-    dump_perf = request.config.getoption("--dump-perf-files-path")
-    # Input sizes
-    B = tkl.sym.B
-    M = tkl.sym.M
-    N = tkl.sym.N
-    K1 = tkl.sym.K1
-    K2 = tkl.sym.K2
-    # Workgroup tile sizes
-    BLOCK_B = tkl.sym.BLOCK_B
-    BLOCK_M = tkl.sym.BLOCK_M
-    BLOCK_N = tkl.sym.BLOCK_N
-    BLOCK_K2 = tkl.sym.BLOCK_K2
-    # Address space (for GPU, shared(1) or global(0))
-    ADDRESS_SPACE = tkl.sym.ADDRESS_SPACE
-    # Other hyperparameters
-    LOAD_ELEMS_PER_THREAD = tkl.sym.LOAD_ELEMS_PER_THREAD
-    STORE_ELEMS_PER_THREAD = tkl.sym.STORE_ELEMS_PER_THREAD
-
-    # Expose user-constraints
-    constraints: list[tkw.Constraint] = [tkw.WorkgroupConstraint(M, BLOCK_M, 0)]
-    constraints += [tkw.WorkgroupConstraint(N, BLOCK_N, 1)]
-    constraints += [tkw.WorkgroupConstraint(B, BLOCK_B, 2)]
-    constraints += [tkw.TilingConstraint(K2, BLOCK_K2)]
-    constraints += [tkw.WaveConstraint(M, BLOCK_M / 2)]
-    constraints += [tkw.WaveConstraint(N, BLOCK_N / 2)]
-
-    constraints += [
-        tkw.HardwareConstraint(
-            threads_per_wave=64,
-            waves_per_block=(2, 2, 1),
-            mma_type=mfma_variant[0],
-            vector_shapes={B: 0},
-        )
-    ]
-
-    i = tkw.IndexMapping.iterator(0)
-    j = tkw.IndexMapping.iterator(1)
-    k = tkw.IndexMapping.iterator(2)
-    mapping = tkw.IndexMapping(
-        num_iterators=3, inputs={B: i, M: j, N: k}, outputs={B: i, N: k, M: j}
-    )
-
-    @tkw.wave(constraints)
-    def chained_gemm_f8(
-        q: tkl.Memory[B, M, K1, GLOBAL_ADDRESS_SPACE, tkl.f16],
-        k: tkl.Memory[B, K2, K1, ADDRESS_SPACE, tkl.f16],
-        v: tkl.Memory[B, N, K2, ADDRESS_SPACE, tkl.f16],
-        c: tkl.Memory[B, N, M, GLOBAL_ADDRESS_SPACE, tkl.f32],
-    ):
-        c_reg = tkl.Register[B, M, N, tkl.f32](0.0)
-
-        @tkw.reduction(K2, init_args=[c_reg])
-        def repeat(
-            acc: tkl.Register[B, M, N, tkl.f32]
-        ) -> tkl.Register[B, M, N, tkl.f32]:
-            inner_acc = tkl.Register[B, K2, M, tkl.f32](0.0)
-            q_reg = tkw.read(q, elements_per_thread=LOAD_ELEMS_PER_THREAD)
-            k_reg = tkw.read(k, elements_per_thread=LOAD_ELEMS_PER_THREAD)
-            q_reg = tkw.cast(q_reg, tkl.f8e4m3fnuz)
-            k_reg = tkw.cast(k_reg, tkl.f8e4m3fnuz)
-            kq_reg = tkw.mma(k_reg, q_reg, inner_acc)
-            qk_reg = tkw.permute(kq_reg, target_shape=[B, M, K2])
-            qk_cast_reg = tkw.cast(qk_reg, tkl.f8e4m3fnuz)
-            v_reg = tkw.read(v, elements_per_thread=LOAD_ELEMS_PER_THREAD)
-            v_reg = tkw.cast(v_reg, tkl.f8e4m3fnuz)
-            acc = tkw.mma(qk_cast_reg, v_reg, acc, mfma_variant[1])
-            return acc
-
-        # repeat represents the results of the loop
-        tkw.write(
-            repeat, c, mapping=mapping, elements_per_thread=STORE_ELEMS_PER_THREAD
-        )
-
-    hyperparams = {
-        ADDRESS_SPACE: SHARED_ADDRESS_SPACE,
-        LOAD_ELEMS_PER_THREAD: get_mfma_load_elems_per_thread(mfma_variant[0]),
-        STORE_ELEMS_PER_THREAD: get_mfma_store_elems_per_thread(mfma_variant[1]),
-        BLOCK_B: 1,
-        BLOCK_M: 64,
-        BLOCK_N: 64,
-        BLOCK_K2: 32,
-        B: shape[0],
-        M: shape[1],
-        N: shape[2],
-        K1: shape[3],
-        K2: shape[4],
-    }
-    hyperparams.update(get_default_scheduling_params())
-    config = get_default_run_config()
-    if run_bench:
-        config["benchmark_batch_size"] = 10
-        config["benchmark_repetitions"] = 3
-    if dump_perf is not None:
-        perf_filename = request.node.name + ".json"
-        config["benchmark_results_file"] = os.path.join(
-            dump_perf, "tk_" + perf_filename
-        )
-
-    with tk.gen.TestLaunchContext(
-        hyperparams,
-        canonicalize=True,
-        run=True,
-        run_bench=run_bench,
-        run_config=config,
-        schedule=enable_scheduling,
-        use_scheduling_barriers=enable_scheduling_barriers,
-    ):
-        torch.manual_seed(0)
-        q = device_randn(shape[0], shape[1], shape[3], dtype=torch.float16)
-        k = device_randn(shape[0], shape[4], shape[3], dtype=torch.float16)
-        v = device_randn(shape[0], shape[2], shape[4], dtype=torch.float16)
-        output = device_zeros(shape[0], shape[2], shape[1], dtype=torch.float32)
-        mb = chained_gemm_f8(q, k, v, output)
-
-        if test_dump_generated_mlir:
-            filename = f"wave_cgemm_{'x'.join(map(str, shape))}.mlir"
-            with open(filename, "w") as f:
-                f.write(mb.module_op.get_asm())
-
-        iree_ref = torch.zeros(shape[0], shape[2], shape[1], dtype=torch.float32)
-        generate_iree_ref(
-            "chain_mmt_f8", [q, k, v], [iree_ref], config, run_bench=run_bench
-        )
-        assert_close(output, iree_ref, atol=7e-5, rtol=2e-3, check_device=False)
-
-
-@require_e2e
-@pytest.mark.parametrize("shape", get_test_shapes("test_attention"))
+@pytest.mark.parametrize("shape", get_test_shapes("attention"))
 @pytest.mark.parametrize("enable_scheduling", [False, True])
 @pytest.mark.parametrize("dynamic_dims", [False, True])
 @pytest.mark.parametrize(
@@ -539,7 +218,7 @@ def testAttention(
             q, k, v, attn_mask=None
         )
 
-        if test_dump_generated_mlir:
+        if dump_generated_mlir:
             filename = f"wave_attention_{'x'.join(map(str, shape))}.mlir"
             with open(filename, "w") as f:
                 f.write(mb.module_op.get_asm())
@@ -548,7 +227,7 @@ def testAttention(
 
 
 @require_e2e
-@pytest.mark.parametrize("shape", get_test_shapes("test_attention"))
+@pytest.mark.parametrize("shape", get_test_shapes("attention"))
 @pytest.mark.parametrize("enable_scheduling", [False])
 @pytest.mark.parametrize("dynamic_dims", [False, True])
 @pytest.mark.parametrize(
@@ -739,7 +418,7 @@ def testAttentionBias(
         a = F.softmax(a, dim=-1)
         torch_ref = torch.matmul(a, v)
 
-        if test_dump_generated_mlir:
+        if dump_generated_mlir:
             filename = f"wave_attention_{'x'.join(map(str, shape))}.mlir"
             with open(filename, "w") as f:
                 f.write(mb.module_op.get_asm())
@@ -753,7 +432,7 @@ def testAttentionBias(
 
 @require_e2e
 @require_cdna3
-@pytest.mark.parametrize("shape", get_test_shapes("test_attention"))
+@pytest.mark.parametrize("shape", get_test_shapes("attention"))
 @pytest.mark.parametrize("enable_scheduling", [False, True])
 @pytest.mark.parametrize(
     "mfma_variant",
@@ -905,100 +584,9 @@ def testAttentionF8(
         torch_ref = torch.nn.functional.scaled_dot_product_attention(
             q, k, v, attn_mask=None
         )
-        if test_dump_generated_mlir:
+        if dump_generated_mlir:
             filename = f"wave_attention_{'x'.join(map(str, shape))}.mlir"
             with open(filename, "w") as f:
                 f.write(mb.module_op.get_asm())
         rmse = torch.sqrt(torch.mean(torch.square(output - torch_ref)))
         assert rmse <= 0.006
-
-
-@require_e2e
-@pytest.mark.parametrize("shape", get_test_shapes("test_attention"))
-@pytest.mark.parametrize("enable_scheduling", [False])
-@pytest.mark.parametrize("dynamic_dims", [False])
-@pytest.mark.parametrize(
-    "mfma_variant",
-    [
-        MMAType.F32_16x16x16_F16,
-    ],
-)
-def testFlashDecoding(
-    shape: tuple[int],
-    enable_scheduling: bool,
-    dynamic_dims: bool,
-    mfma_variant: MMAType,
-    request,
-):
-    run_bench = request.config.getoption("--runperf")
-    dump_perf = request.config.getoption("--dump-perf-files-path")
-    phase_0, phase_1, hyperparams_0, hyperparams_1 = get_decode_attention_kernels(
-        shape, mfma_variant
-    )
-    hyperparams_0.update(get_default_scheduling_params())
-    hyperparams_1.update(get_default_scheduling_params())
-    config = get_default_run_config()
-    if run_bench:
-        config["benchmark_batch_size"] = 10
-        config["benchmark_repetitions"] = 3
-    if dump_perf is not None:
-        perf_filename = request.node.name + ".json"
-        config["benchmark_results_file"] = os.path.join(
-            dump_perf, "tk_" + perf_filename
-        )
-
-    torch.manual_seed(0)
-    B, M, N, K1, K2 = shape
-    U = hyperparams_0[index_symbol("U")]
-    q = device_randn(B, M, K1, dtype=torch.float16)
-    k = device_randn(B, K2, K1, dtype=torch.float16)
-    v = device_randn(B, K2, N, dtype=torch.float16)
-    phase_0_output = device_zeros(U, B, N, M, dtype=torch.float32)
-    phase_0_output_max = device_zeros(U, B, M, dtype=torch.float32)
-    output = device_zeros(B, M, N, dtype=torch.float32)
-    log2e = 1.44269504089
-    dk_sqrt = math.sqrt(1.0 / K1)
-
-    with tk.gen.TestLaunchContext(
-        hyperparams_0,
-        canonicalize=True,
-        run=True,
-        run_bench=run_bench,
-        run_config=config,
-        schedule=enable_scheduling,
-        use_scheduling_barriers=enable_scheduling_barriers,
-    ):
-        # TODO: Add scaling of QK as part of kernel.
-        mb_qk = phase_0(
-            q * dk_sqrt * log2e,
-            k,
-            v.permute([0, 2, 1]),
-            phase_0_output,
-            phase_0_output_max,
-        )
-
-    with tk.gen.TestLaunchContext(
-        hyperparams_1,
-        canonicalize=True,
-        run=True,
-        run_bench=run_bench,
-        run_config=config,
-        schedule=enable_scheduling,
-        use_scheduling_barriers=enable_scheduling_barriers,
-    ):
-        # TODO: Add variant of non-transposed V attention kernel.
-        mb_sv = phase_1(phase_0_output, phase_0_output_max, output)
-
-    torch_ref = torch.nn.functional.scaled_dot_product_attention(
-        q, k, v, attn_mask=None
-    )
-
-    if test_dump_generated_mlir:
-        filename = f"wave_phase_0_kernel_{'x'.join(map(str, shape))}.mlir"
-        with open(filename, "w") as f:
-            f.write(mb_qk.module_op.get_asm())
-        filename = f"wave_phase_1_kernel_{'x'.join(map(str, shape))}.mlir"
-        with open(filename, "w") as f:
-            f.write(mb_sv.module_op.get_asm())
-
-    assert_allclose(output, torch_ref)
